# openapi/karma-v1.yaml
openapi: 3.1.0
info:
  title: Karma Ecosystem API
  version: 1.0.0
  summary: Payment intent, evidence verification, and risk alert APIs.
servers:
  - url: http://127.0.0.1:8811
security:
  - bearerAuth: []
paths:
...
  /v1/capacity/{identityId}:
    get:
      summary: Get identity capacity ledger
      operationId: getCapacity
      parameters:
        - in: path
          name: identityId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Capacity snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapacityState'
  /v1/capacity/{identityId}/lock:
    post:
      summary: Lock USDC and mint 1:1 bill credits
      operationId: lockCapacity
      parameters:
        - in: path
          name: identityId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmountRequest'
      responses:
        '200':
          description: Updated capacity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapacityState'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/capacity/{identityId}/release:
    post:
      summary: Release unused available credits
      operationId: releaseCapacity
      parameters:
        - in: path
          name: identityId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmountRequest'
      responses:
        '200':
          description: Updated capacity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapacityState'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/BadRequest'
...
  /v1/progress:
    post:
      summary: Submit progress receipt
      operationId: submitProgressReceipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressReceipt'
      responses:
        '201':
          description: Progress receipt submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressReceipt'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/BadRequest'
  /v1/progress/{progressReceiptId}/confirm:
    post:
      summary: Confirm progress receipt
      operationId: confirmProgressReceipt
      parameters:
        - in: path
          name: progressReceiptId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Confirmed progress receipt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressReceipt'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/BadRequest'
  /v1/progress/task/{taskId}:
    get:
      summary: List progress receipts for task
      operationId: listProgressReceipts
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Progress receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProgressReceipt'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AmountRequest:
      type: object
      properties:
        amount:
          type: number
          exclusiveMinimum: 0
      required: [amount]
    CapacityState:
      type: object
      properties:
        identity_id:
          type: string
        total_locked_usdc:
          type: number
        total_bill_credits:
          type: number
        available_credits:
          type: number
        reserved_credits:
          type: number
        in_progress_credits:
          type: number
        confirmed_progress_credits:
          type: number
        disputed_credits:
          type: number
        pending_settlement_credits:
          type: number
        burned_credits:
          type: number
        released_credits:
          type: number
        updated_at:
          type: string
          format: date-time
      required:
        - identity_id
        - total_locked_usdc
        - total_bill_credits
        - available_credits
        - reserved_credits
        - in_progress_credits
        - confirmed_progress_credits
        - disputed_credits
        - pending_settlement_credits
        - burned_credits
        - released_credits
        - updated_at
...
    ProgressReceipt:
      type: object
      properties:
        progress_receipt_id:
          type: string
        task_id:
          type: string
        seller_identity_id:
          type: string
        progress_percent:
          type: number
          minimum: 0
          maximum: 100
        claimed_value_percent:
          type: number
          minimum: 0
          maximum: 100
        evidence_hash:
          type: string
        runtime_log_hash:
          type: string
        timestamp:
          type: string
          format: date-time
        seller_signature:
          type: string
        validation_method:
          type: string
        confirmation_status:
          type: string
          enum: [pending, confirmed, rejected]
        confirmed_at:
          type: string
          format: date-time
      required:
        - progress_receipt_id
        - task_id
        - seller_identity_id
        - progress_percent
        - claimed_value_percent
        - evidence_hash
        - runtime_log_hash
        - timestamp
        - seller_signature
        - validation_method
        - confirmation_status
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
          required: [code, message, requestId]
      required: [error]
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'